requirements.txt
# SPDX-License-Identifier: FSFAP
# Copyright (C) 2020 Michael Zhang
# Copyright (C) 2020-2024 Colin B. Macdonald
# Copyright (C) 2020 Vala Vakilian
# Copyright (C) 2022 Elizabeth Xiao
# Copyright (C) 2022-2023 Andrew Rechnitzer
# Copyright (C) 2022 Brennen Chiu
# Copyright (C) 2022-2024 Edith Coates
# Copyright (C) 2023 Julian Lapenna
# Copyright (C) 2023 Natalie Balashov
# Copyright (C) 2024 Bryan Tanady

# Copying and distribution of this file, with or without modification,
# are permitted in any medium without royalty provided the copyright
# notice and this notice are preserved.  This file is offered as-is,
# without any warranty.

# Not server deps
platformdirs==4.2.2
PyQt6==6.6.1
PyQt6-Qt6==6.6.3
PyQt6-sip==13.6.0
pyspellchecker==0.8.1
pytest-qt==4.4.0

# Legacy-server only deps
aiohttp==3.9.5
cryptography==42.0.8
peewee==3.17.6

arrow==1.3.0
beautifulsoup4==4.12.3
canvasapi==3.2.0
debugpy==1.8.2
django==4.2.14
django-braces==1.15.0
django-filter==24.2
django-htmx==1.18.0
django-huey==1.2.1
django-humanize==0.1.2
django-mathfilters==1.0.0
django-polymorphic==3.1.0
django-reset-migrations==0.4.0
django-session-timeout==0.1.0
djangorestframework==3.15.2
exif==1.6.0
file-magic==0.4.1
fonttools==4.53.1
gunicorn==22.0.0
importlib-resources==6.4.0
imutils==0.5.4
matplotlib==3.8.4
model-bakery==1.18.2
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.1
pandas==2.1.4
passlib==1.7.4
Pillow==10.4.0
psycopg2-binary==2.9.9
pymupdf==1.24.7
PyMySQL==1.1.1
random-username==1.0.2
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
segno==1.6.1
stdiomask==0.0.6
tabulate==0.9.0
tomli==2.0.1
tomlkit==0.13.0
tqdm==4.66.4
urllib3==2.2.2
weasyprint==60.2
whitenoise==6.7.0
zipfly==6.0.5
zxing-cpp==2.2.0

#---------- for dev / debug work - comment out unless using it.
# nplusone==1.0.0  # for hunting down potential n+1 query problems
